# backend/src/main/resources/static/docs/openapi.yaml
openapi: 3.0.1
info:
  title: DraftURL API
  description: DraftURL 백엔드 REST API 문서
  version: 1.0.0
servers:
- url: http://localhost:8080
...
paths:
...
  /api/v1/auth/logout:
    post:
      tags:
      - Auth API
      summary: 로그아웃
      description: Refresh Token을 무효화하여 로그아웃을 처리합니다.
      operationId: auth-logout
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/api-v1-auth-logout913878229"
            examples:
              auth-logout:
                value: |-
                  {
                    "refreshToken" : "rt_abcdef123456"
                  }
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api-v1-auth-logout-1308105853"
              examples:
                auth-logout:
                  value: |-
                    {
                      "success" : true,
                      "data" : null,
                      "error" : null
                    }
...
  /api/v1/auth/refresh:
    post:
      tags:
      - Auth API
      summary: 토큰 갱신
      description: Refresh Token을 사용하여 새로운 Access Token과 Refresh Token을 발급합니다. Refresh
        Token Rotation이 적용되어 기존 Refresh Token은 무효화됩니다.
      operationId: auth-refresh
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/api-v1-auth-refresh1051830133"
            examples:
              auth-refresh:
                value: |-
                  {
                    "refreshToken" : "rt_abcdef123456"
                  }
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api-v1-auth-refresh-141030569"
              examples:
                auth-refresh:
                  value: |-
                    {
                      "success" : true,
                      "data" : {
                        "accessToken" : "eyJhbGciOiJIUzI1NiJ9.new_access_token.example",
                        "refreshToken" : "rt_newtoken789",
                        "expiresIn" : 3600,
                        "user" : {
                          "id" : "00000000-0000-0000-0000-000000000001",
                          "email" : "user@example.com",
                          "name" : "홍길동",
                          "profileImage" : "https://example.com/avatar.jpg",
                          "plan" : "free"
                        }
                      },
                      "error" : null
                    }
...
  /api/v1/auth/oauth2/callback/{provider}:
    post:
      tags:
      - Auth API
      summary: OAuth2 콜백 처리
      description: "OAuth2 인가 코드를 사용하여 로그인 또는 회원가입을 처리하고 JWT 토큰을 발급합니다. 지원 Provider:\
        \ github, google"
      operationId: auth-oauth-callback
      parameters:
      - name: provider
        in: path
        description: "OAuth2 Provider (github, google)"
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/api-v1-auth-oauth2-callback-provider1241930397"
            examples:
              auth-oauth-callback:
                value: |-
                  {
                    "code" : "auth_code_abc123",
                    "redirectUri" : "http://localhost:3000/auth/callback",
                    "state" : "random_state_value"
                  }
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api-v1-auth-oauth2-callback-provider-321854121"
              examples:
                auth-oauth-callback:
                  value: |-
                    {
                      "success" : true,
                      "data" : {
                        "accessToken" : "eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiIxIiwidXNlcklkIjoiMDAwMDAwMDAifQ.example",
                        "refreshToken" : "rt_abcdef123456",
                        "expiresIn" : 3600,
                        "user" : {
                          "id" : "00000000-0000-0000-0000-000000000001",
                          "email" : "user@example.com",
                          "name" : "홍길동",
                          "profileImage" : "https://example.com/avatar.jpg",
                          "plan" : "free"
                        }
                      },
                      "error" : null
                    }
components:
  schemas:
    api-v1-auth-refresh1051830133:
      required:
      - refreshToken
      type: object
      properties:
        refreshToken:
          type: string
          description: Refresh Token
    api-v1-auth-logout-1308105853:
      required:
      - data
      - error
      - success
      type: object
      properties:
        success:
          type: boolean
          description: 요청 성공 여부
...
    api-v1-auth-oauth2-callback-provider-321854121:
      required:
      - error
      - success
      type: object
      properties:
        data:
          required:
          - accessToken
          - expiresIn
          - refreshToken
          type: object
          properties:
            expiresIn:
              type: number
              description: Access Token 만료 시간 (초)
            accessToken:
              type: string
              description: JWT Access Token
            user:
              required:
              - email
              - id
              - name
              - plan
              - profileImage
              type: object
              properties:
                name:
                  type: string
                  description: 이름
                profileImage:
                  type: string
                  description: 프로필 이미지 URL
                id:
                  type: string
                  description: 사용자 ID
                plan:
                  type: string
                  description: "요금제 (free, pro)"
                email:
                  type: string
                  description: 이메일
            refreshToken:
              type: string
              description: Refresh Token
        success:
          type: boolean
          description: 요청 성공 여부
...
    api-v1-auth-logout913878229:
      required:
      - refreshToken
      type: object
      properties:
        refreshToken:
          type: string
          description: 무효화할 Refresh Token
    api-v1-auth-refresh-141030569:
      required:
      - error
      - success
      type: object
      properties:
        data:
          required:
          - accessToken
          - expiresIn
          - refreshToken
          type: object
          properties:
            expiresIn:
              type: number
              description: Access Token 만료 시간 (초)
            accessToken:
              type: string
              description: 새로운 JWT Access Token
            user:
              required:
              - email
              - id
              - name
              - plan
              - profileImage
              type: object
              properties:
                name:
                  type: string
                  description: 이름
                profileImage:
                  type: string
                  description: 프로필 이미지 URL
                id:
                  type: string
                  description: 사용자 ID
                plan:
                  type: string
                  description: "요금제 (free, pro)"
                email:
                  type: string
                  description: 이메일
            refreshToken:
              type: string
              description: 새로운 Refresh Token
        success:
          type: boolean
          description: 요청 성공 여부
...
    api-v1-auth-oauth2-callback-provider1241930397:
      required:
      - code
      - redirectUri
      - state
      type: object
      properties:
        redirectUri:
          type: string
          description: 리다이렉트 URI
        code:
          type: string
          description: OAuth2 인가 코드
        state:
          type: string
          description: CSRF 방지를 위한 state 값
...